Added create user function.
Added a api call to create user.
The info from the input files is used to create the user.
Also adde obscureText to the password input fileds so you cant see the password writen in.
Added a check password function, that checks if the password and confirmpassword are the same.