Added method to fetch lists
Added API point that fetches lists in my list page.
Also made it so that when the my list button is prest the method is run and shows the list it gets.
For this to work i added a model file cald StockList. Named it StockList so it wont be conflict from List that allrady exists.